requirements.txt
aiohttp
pyyaml
loguru